Fix dialyzer, eunit issues
Redbug was using an invalid version of net_kernel:start/1 passing a map.  This does not appear to be supported, and dialyzer rightly complains - from OTP 24 start/2 should be used, but start/1 only supports a list not a map.

This appears to function (somehow) in OTP 26, in that the redbug_dist_eunit tests pass ??!!.  However, those tests fail in OTP 24.

Change here makes dialyzer happy for both OTP 24 and OTP 26 (note that the hidden option wasn't introduced until OTP 25, so start/2 cannot be used in this way in OTP 24).  However, although eunit tests pass now for OTP 26 and OTP 25 - they still fail for OTP 24.

This commit is an improvement - in that dialyzer is now OK, and makes the eunit test situtation no worse.

the spec for redbug:start/2 is also removed.  This is incorrect in that the output does not appear to be a tuple of two integers (the tests expect a 3-tuple for example).   Also the test use a list not a map for options (and as the map is simply converted to a list, this seems OK).